:bug: Fix resize deleting history
Shrinking the terminal would delete history. Changed the resize function
to behave more like a real terminal. If the terminal is resized to be
smaller, take a display row off the top while keeping the line in
history. If the terminal is resized to be larger, add a display row to
the top.