Remove scroll to bottom on write
Now the terminal will only scroll to the bottom if the user is at the
last line. This way if the user scrolls up then they won’t be forced
back down to the last line. Auto scrolling can be enabled again once
the user scrolls back down to the last line.