:bug: Return focus on refresh
Refresh method causes the terminal to lose focus because it overwrites
the terminal object in the HTML document. This momentarily causes the
terminal to not exist and therefore lose focus (document.activeElement).
Make sure we return focus after appending to the parent element if the
terminal should be in focus.