:bug: Fix textarea bugs
Focus on textarea only if changing the currently focused terminal. Focus
on textarea on refresh. Prevent clearSelection from accessing an index
from the history array that is undefined. Disable the implemented
selection in favor of the default selection provided by the browser.
Might use the implemented selection later to improve copy-paste. Focus
on textarea in keypress event if it is not already in focus.